+++ src/providers/meld.rs
@@ -15,6 +15,7 @@ use {
     reqwest::StatusCode,
     serde::{Deserialize, Serialize},
     std::{sync::Arc, time::SystemTime},
+    tokio::task::JoinSet,
     tracing::log::error,
     url::Url,
 };
@@ -22,6 +23,7 @@ use {
 const API_VERSION: &str = "2023-12-19";
 const DEFAULT_CATEGORY: &str = "CRYPTO_ONRAMP";
 const DEFAULT_SESSION_TYPE: &str = "BUY";
+const EMPTY_QUOTES_ERROR_CODE: &str = "NO_VALID_QUOTES";
 
 #[derive(Debug)]
 pub struct MeldProvider {
@@ -66,6 +68,75 @@ impl MeldProvider {
             .send()
             .await
     }
+
+    /// Fetches quotes for a single payment type
+    async fn fetch_quotes_for_payment_type(
+        payment_type: String,
+        mut params: MultiQuotesQueryParams,
+        url: Url,
+        metrics: Arc<Metrics>,
+        http_client: reqwest::Client,
+        api_key: String,
+    ) -> RpcResult<Vec<QuotesResponse>> {
+        params.payment_method_type = Some(payment_type);
+
+        let latency_start = SystemTime::now();
+        let response = http_client
+            .post(url)
+            .json(&params)
+            .header("Meld-Version", API_VERSION)
+            .header("Authorization", format!("BASIC {}", api_key))
+            .send()
+            .await
+            .map_err(|e| {
+                error!("Error sending request to Meld get quotes: {:?}", e);
+                RpcError::OnRampProviderError
+            })?;
+        metrics.add_latency_and_status_code_for_provider(
+            ProviderKind::Meld,
+            response.status().into(),
+            latency_start,
+            None,
+            Some("onramp_multi_quotes".to_string()),
+        );
+
+        if !response.status().is_success() {
+            // Passing through error description for the error context
+            // if user parameter is invalid (got 400 status code from the provider)
+            if matches!(
+                response.status(),
+                StatusCode::BAD_REQUEST | StatusCode::UNPROCESSABLE_ENTITY
+            ) {
+                let response_error = match response.json::<MeldErrorResponse>().await {
+                    Ok(response_error) => response_error,
+                    Err(e) => {
+                        error!(
+                            "Error parsing Meld HTTP 400 Bad Request error response {:?}",
+                            e
+                        );
+                        // Respond to the client with a generic error message and HTTP 400 anyway
+                        MeldErrorResponse {
+                            code: "UNKNOWN".to_string(),
+                            message: "Invalid parameter".to_string(),
+                        }
+                    }
+                };
+                return Err(RpcError::ConversionInvalidParameterWithCode(
+                    response_error.code,
+                    response_error.message,
+                ));
+            }
+
+            error!(
+                "Error on Meld get quotes. Status is not OK: {:?}",
+                response.status(),
+            );
+            return Err(RpcError::OnRampProviderError);
+        }
+
+        let response_quotes = response.json::<MeldQuotesResponse>().await?;
+        Ok(response_quotes.quotes)
+    }
 }
 
 #[derive(Debug, Serialize, Clone)]
@@ -81,11 +152,21 @@ pub struct MeldQuotesResponse {
     pub quotes: Vec<QuotesResponse>,
 }
 
+// Can be used for both old (Meld prod) and new (Meld sandbox) error response schemas
 #[derive(Debug, Serialize, Deserialize, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct MeldErrorResponse {
+    #[serde(alias = "error")]
     pub code: String,
     pub message: String,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct PaymentMethod {
+    pub name: String,
+    pub payment_method: String,
+    pub payment_type: String,
 }
 
 #[async_trait]
@@ -177,6 +258,10 @@ impl OnRampMultiProvider for MeldProvider {
                 "{}/service-providers/limits/fiat-currency-purchases",
                 self.api_base_url
             ),
+            PropertyType::CountriesDefaults => format!(
+                "{}/service-providers/properties/defaults/by-country",
+                self.api_base_url
+            ),
         };
         let mut url = Url::parse(&base_url).map_err(|_| RpcError::OnRampParseURLError)?;
         if let Some(countries) = params.countries {
@@ -186,7 +271,13 @@ impl OnRampMultiProvider for MeldProvider {
             .append_pair("categories", DEFAULT_CATEGORY);
 
         let latency_start = SystemTime::now();
-        let response = self.send_get_request(url).await?;
+        let response = self.send_get_request(url).await.map_err(|e| {
+            error!(
+                "Error sending request to Meld providers properties: {:?}",
+                e
+            );
+            RpcError::OnRampProviderError
+        })?;
         metrics.add_latency_and_status_code_for_provider(
             self.provider_kind,
             response.status().into(),
@@ -243,7 +334,11 @@ impl OnRampMultiProvider for MeldProvider {
                     session_data: params.session_data,
                 },
             )
-            .await?;
+            .await
+            .map_err(|e| {
+                error!("Error sending request to Meld get widget: {:?}", e);
+                RpcError::OnRampProviderError
+            })?;
         metrics.add_latency_and_status_code_for_provider(
             self.provider_kind,
             response.status().into(),
@@ -290,52 +385,91 @@ impl OnRampMultiProvider for MeldProvider {
         let base = format!("{}/payments/crypto/quote", self.api_base_url);
         let url = Url::parse(&base).map_err(|_| RpcError::OnRampParseURLError)?;
 
-        let latency_start = SystemTime::now();
-        let response = self.send_post_request(url, &params).await?;
-        metrics.add_latency_and_status_code_for_provider(
-            self.provider_kind,
-            response.status().into(),
-            latency_start,
-            None,
-            Some("onramp_multi_quotes".to_string()),
-        );
-
-        if !response.status().is_success() {
-            // Passing through error description for the error context
-            // if user parameter is invalid (got 400 status code from the provider)
-            if matches!(
-                response.status(),
-                StatusCode::BAD_REQUEST | StatusCode::UNPROCESSABLE_ENTITY
-            ) {
-                let response_error = match response.json::<MeldErrorResponse>().await {
-                    Ok(response_error) => response_error,
-                    Err(e) => {
-                        error!(
-                            "Error parsing Meld HTTP 400 Bad Request error response {:?}",
-                            e
-                        );
-                        // Respond to the client with a generic error message and HTTP 400 anyway
-                        MeldErrorResponse {
-                            code: "BAD_REQUEST".to_string(),
-                            message: "Invalid parameter".to_string(),
-                        }
-                    }
-                };
-                return Err(RpcError::ConversionInvalidParameterWithCode(
-                    response_error.code,
-                    response_error.message,
-                ));
+        // Get all payment-types based on the country if not payment type was provided.
+        // Country is optional, but we are providing it anyways and it's required
+        // to get the payment types we will use the fallback to the US as a default.
+        let mut payment_types: Vec<String> = Vec::new();
+        if let Some(payment_type) = &params.payment_method_type {
+            payment_types.push(payment_type.to_string());
+        } else {
+            let country = params.clone().country_code.unwrap_or("US".to_string());
+            let providers_properties_response = self
+                .get_providers_properties(
+                    ProvidersPropertiesQueryParams {
+                        project_id: params.project_id.clone(),
+                        r#type: PropertyType::PaymentMethods,
+                        countries: Some(country.to_string()),
+                    },
+                    metrics.clone(),
+                )
+                .await?;
+            let payment_methods: Vec<PaymentMethod> =
+                serde_json::from_value(providers_properties_response)?;
+            for payment_method in payment_methods {
+                // Push only if the payment type is not already in the list
+                // because Meld response contains duplicates for the same payment type
+                // but with different payment methods.
+                if !payment_types.contains(&payment_method.payment_type) {
+                    payment_types.push(payment_method.payment_type);
+                }
             }
+        };
 
-            error!(
-                "Error on Meld get quotes. Status is not OK: {:?}",
-                response.status(),
-            );
-            return Err(RpcError::OnRampProviderError);
+        // We are not expecting more than 20 payment types, we should stop
+        // if we have more than 20 payment types.
+        if payment_types.len() > 20 {
+            return Err(RpcError::ConversionProviderInternalError(
+                "Too many payment types".to_string(),
+            ));
         }
 
-        let response_quotes = response.json::<MeldQuotesResponse>().await?;
+        // Get quotes for each payment type in parallel and aggregate results
+        // otherwise only the card payment is provided in quotes if there are no
+        // payment type was provided to the request, but we want to get all
+        // available quotes for all payment types.
+        let mut join_set = JoinSet::new();
+        for payment_type in payment_types {
+            let params = params.clone();
+            let url = url.clone();
+            let metrics = metrics.clone();
+            let http_client = self.http_client.clone();
+            let api_key = self.api_key.clone();
 
-        Ok(response_quotes.quotes)
+            join_set.spawn(async move {
+                Self::fetch_quotes_for_payment_type(
+                    payment_type,
+                    params,
+                    url,
+                    metrics,
+                    http_client,
+                    api_key,
+                )
+                .await
+            });
+        }
+
+        let mut quotes = Vec::new();
+        while let Some(result) = join_set.join_next().await {
+            match result {
+                Ok(Ok(quotes_response)) => quotes.extend(quotes_response),
+                Ok(Err(e)) => {
+                    // Check if this is an EMPTY_QUOTES_ERROR_CODE error, if so continue to next payment type
+                    // because the list can be fulfilled with quotes for other payment types.
+                    if let RpcError::ConversionInvalidParameterWithCode(code, _) = &e {
+                        if code == EMPTY_QUOTES_ERROR_CODE {
+                            error!("No valid quotes for payment type, continuing to next");
+                            continue;
+                        }
+                    }
+                    return Err(e);
+                }
+                Err(e) => {
+                    error!("Error on getting Meld quotes in parallel: {:?}", e);
+                    return Err(RpcError::OnRampProviderError);
+                }
+            }
+        }
+
+        Ok(quotes)
     }
 }