+++ src/providers/mod.rs
@@ -68,6 +68,7 @@ mod arbitrum;
 mod aurora;
 mod base;
 mod binance;
+mod blast;
 mod bungee;
 mod callstatic;
 mod coinbase;
@@ -84,11 +85,11 @@ mod morph;
 mod near;
 mod odyssey;
 mod one_inch;
-mod onerpc;
 mod pimlico;
 mod pokt;
 mod publicnode;
 mod quicknode;
+mod rootstock;
 mod solscan;
 mod sui;
 mod syndica;
@@ -107,6 +108,7 @@ pub use {
     aurora::AuroraProvider,
     base::BaseProvider,
     binance::BinanceProvider,
+    blast::BlastProvider,
     bungee::BungeeProvider,
     callstatic::CallStaticProvider,
     drpc::DrpcProvider,
@@ -121,11 +123,11 @@ pub use {
     near::NearProvider,
     odyssey::OdysseyProvider,
     one_inch::OneInchProvider,
-    onerpc::OneRpcProvider,
     pimlico::PimlicoProvider,
     pokt::PoktProvider,
     publicnode::PublicnodeProvider,
     quicknode::QuicknodeProvider,
+    rootstock::RootstockProvider,
     solscan::SolScanProvider,
     sui::SuiProvider,
     syndica::{SyndicaProvider, SyndicaWsProvider},
@@ -183,6 +185,8 @@ pub struct ProvidersConfig {
     pub meld_api_url: String,
     /// CallStatic API key
     pub callstatic_api_key: String,
+    /// Blast.io API key
+    pub blast_api_key: String,
 
     pub override_bundler_urls: Option<MockAltoUrls>,
 }
@@ -745,10 +749,11 @@ pub enum ProviderKind {
     Sui,
     Hiro,
     CallStatic,
-    OneRpc,
     TheRpc,
     Moonbeam,
     Edexa,
+    Blast,
+    Rootstock,
 }
 
 impl Display for ProviderKind {
@@ -787,10 +792,11 @@ impl Display for ProviderKind {
                 ProviderKind::Sui => "Sui",
                 ProviderKind::Hiro => "Hiro",
                 ProviderKind::CallStatic => "CallStatic",
-                ProviderKind::OneRpc => "OneRpc",
                 ProviderKind::TheRpc => "TheRpc",
                 ProviderKind::Moonbeam => "Moonbeam",
                 ProviderKind::Edexa => "edeXa",
+                ProviderKind::Blast => "Blast",
+                ProviderKind::Rootstock => "Rootstock",
             }
         )
     }
@@ -830,10 +836,11 @@ impl ProviderKind {
             "Sui" => Some(Self::Sui),
             "Hiro" => Some(Self::Hiro),
             "CallStatic" => Some(Self::CallStatic),
-            "OneRpc" => Some(Self::OneRpc),
             "TheRpc" => Some(Self::TheRpc),
             "Moonbeam" => Some(Self::Moonbeam),
             "edeXa" => Some(Self::Edexa),
+            "Blast" => Some(Self::Blast),
+            "Rootstock" => Some(Self::Rootstock),
             _ => None,
         }
     }